update dhcpcd to v6.8.2, based on https://android.googlesource.com/platform/external/dhcpcd-6.8.2/
commit info:
commit  a34a61b23d287c7d24a2d34c7f8c191641747316        [log] [tgz]
author  Dan Willemsen <dwillemsen@google.com>   Mon Mar 12 22:04:15 2018
+0000
committer       android-build-merger <android-build-merger@google.com>
Mon Mar 12 22:04:15 2018 +0000
tree    472ee9502ffd3f3a453ac1adcc425ba1bfb37139
parent  b1e043e5fd15a8e93beb9dbe85100dddb62863eb [diff]
parent  87b492fdcd1e94e3e1eb7c0396d6cb4a4282e443 [diff]

+++ compat/strlcpy.c
@@ -1,4 +1,4 @@
-/* 
+/*
  * dhcpcd - DHCP client daemon
  * Copyright (c) 2006-2009 Roy Marples <roy@marples.name>
  * All rights reserved
@@ -47,5 +47,6 @@ strlcpy(char *dst, const char *src, size_t size)
 		while (*src++);
 	}
 
-	return src - s - 1;
+
+	return (size_t)(src - s - 1);
 }
